+++ src/modules/Organization/index.ts
@@ -1,4 +1,4 @@
-import Organization from './Pages/Organization'
+import OrganizationPage from './Pages/OrganizationPage.tsx'
 import LeavePolicy from '../Leaves/Pages/LeavePolicy.tsx'
 import OfficialHolidays from './Pages/OfficialHolidays'
 import Teams from './Pages/Teams'
@@ -7,4 +7,4 @@ import UpdateTeam from './Pages/UpdateTeam'
 import ImportHolidays from './Pages/ImportHolidays'
 import Leaves from '../Leaves/Pages/Leaves.tsx'
 
-export {LeavePolicy, Organization, OfficialHolidays, Teams, CreateTeam, UpdateTeam, ImportHolidays, Leaves }
+export {LeavePolicy, OrganizationPage, OfficialHolidays, Teams, CreateTeam, UpdateTeam, ImportHolidays, Leaves }
